fixes deployment error
