+++ frontend/src/App.jsx
@@ -22,10 +22,10 @@ import {
   SingleMedicine,
   Medicines,
   GoToTop,
-  OrderedProduct,
-  Bot
+  CartPage,
+  Bot,
 } from "./import-export/ImportExport.js";
-import TopSpecialitiesCard from "./components/TopSpecialitiesCard.jsx";
+// import TopSpecialitiesCard from "./components/TopSpecialitiesCard.jsx";
 import SpecialitiesPage from "./pages/SpecialitiesPage.jsx";
 
 function App() {
@@ -46,11 +46,11 @@ function App() {
           <Route path="/medicine-cart" element={<AddtoCart />} />
           <Route path="/login" element={<Login />} />
           <Route path="/signup" element={<Signup />} />
-          <Route path="/specialities" element={<SpecialitiesPage/>}/>
+          <Route path="/specialities" element={<SpecialitiesPage />} />
+          <Route path="/cart" element={<CartPage />} />
           <Route path="/*" element={<ErrorPage />} />
-          <Route path="/order" element={<OrderedProduct/>}/>
         </Routes>
-        <Bot/>
+        <Bot />
         <GoToTop />
         <Footer />
         <ToastContainer position="top-center" />
